Move to an API approach for dashboards
Expose available API endpoints to the frontend
