+++ tests/unit/lms/views/dashboard/api/assignment_test.py
@@ -1,0 +1,91 @@
+from unittest.mock import sentinel
+
+import pytest
+
+from lms.views.dashboard.api.assignment import AssignmentViews
+from tests import factories
+
+pytestmark = pytest.mark.usefixtures("h_api", "assignment_service")
+
+
+class TestAssignmentViews:
+    def test_assignment(self, views, pyramid_request, assignment_service):
+        pyramid_request.matchdict["id_"] = sentinel.id
+        assignment = factories.Assignment()
+        assignment_service.get_by_id.return_value = assignment
+
+        response = views.assignment()
+
+        assignment_service.get_by_id.assert_called_once_with(sentinel.id)
+
+        assert response == {
+            "id": assignment.id,
+            "title": assignment.title,
+        }
+
+    def test_assignment_stats(self, views, pyramid_request, assignment_service, h_api):
+        # User returned by the stats endpoint
+        student = factories.User()
+        # User with no annotations
+        student_no_annos = factories.User(display_name="Homer")
+        # User with no annotations and no name
+        student_no_annos_no_name = factories.User(display_name=None)
+
+        pyramid_request.matchdict["id_"] = sentinel.id
+        assignment = factories.Assignment()
+        assignment_service.get_members.return_value = [
+            student,
+            student_no_annos,
+            student_no_annos_no_name,
+        ]
+        assignment_service.get_by_id.return_value = assignment
+        stats = [
+            {
+                "display_name": student.display_name,
+                "annotations": sentinel.annotations,
+                "replies": sentinel.replies,
+                "userid": student.h_userid,
+                "last_activity": sentinel.last_activity,
+            },
+            {
+                "display_name": sentinel.display_name,
+                "annotations": sentinel.annotations,
+                "replies": sentinel.replies,
+                "userid": "TEACHER",
+                "last_activity": sentinel.last_activity,
+            },
+        ]
+
+        h_api.get_assignment_stats.return_value = stats
+
+        response = views.assignment_stats()
+
+        assignment_service.get_by_id.assert_called_once_with(sentinel.id)
+        h_api.get_assignment_stats.assert_called_once_with(
+            [g.authority_provided_id for g in assignment.groupings],
+            assignment.resource_link_id,
+        )
+        assert response == [
+            {
+                "display_name": student.display_name,
+                "annotations": sentinel.annotations,
+                "replies": sentinel.replies,
+                "last_activity": sentinel.last_activity,
+            },
+            {
+                "display_name": student_no_annos.display_name,
+                "annotations": 0,
+                "replies": 0,
+                "last_activity": None,
+            },
+            {
+                "display_name": f"Student {student_no_annos_no_name.user_id[:10]}",
+                "annotations": 0,
+                "replies": 0,
+                "last_activity": None,
+            },
+        ]
+
+    @pytest.fixture
+    def views(self, pyramid_request):
+        return AssignmentViews(pyramid_request)
